food_backend/requirements.txt
amqp==5.2.0
aniso8601==7.0.0
APScheduler==3.10.4
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
kombu==5.3.7
pillow==10.4.0
promise==2.3
prompt_toolkit==3.0.46
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.5
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
